+++ naga/src/back/hlsl/conv.rs
@@ -162,6 +162,7 @@ impl crate::BuiltIn {
             Self::FrontFacing => "SV_IsFrontFace",
             Self::PrimitiveIndex => "SV_PrimitiveID",
             Self::Barycentric => "SV_Barycentrics",
+            Self::BarycentricNoPerspective => "SV_Barycentrics",
             Self::SampleIndex => "SV_SampleIndex",
             Self::SampleMask => "SV_Coverage",
             // compute