+++ naga/src/ir/mod.rs
@@ -414,6 +414,8 @@ pub enum BuiltIn {
     /// Read in fragment shaders
     Barycentric,
     /// Read in fragment shaders
+    BarycentricNoPerspective,
+    /// Read in fragment shaders
     SampleIndex,
     /// Read or written in fragment shaders
     SampleMask,