+++ wgpu-core/src/validation/shader_io_deductions.rs
@@ -48,7 +48,7 @@ impl MaxFragmentShaderInputDeduction {
                 | InterStageBuiltIn::SubgroupSize
                 | InterStageBuiltIn::ViewIndex
                 | InterStageBuiltIn::PointCoord => 1,
-                InterStageBuiltIn::Barycentric => 3,
+                InterStageBuiltIn::Barycentric | InterStageBuiltIn::BarycentricNoPerspective => 3,
                 InterStageBuiltIn::Position => 0,
             },
         }
@@ -68,6 +68,7 @@ impl MaxFragmentShaderInputDeduction {
 
             BuiltIn::PointCoord => InterStageBuiltIn::PointCoord,
             BuiltIn::Barycentric => InterStageBuiltIn::Barycentric,
+            BuiltIn::BarycentricNoPerspective => InterStageBuiltIn::BarycentricNoPerspective,
             BuiltIn::ViewIndex => InterStageBuiltIn::ViewIndex,
 
             BuiltIn::BaseInstance
@@ -118,6 +119,7 @@ pub enum InterStageBuiltIn {
     // Non-standard
     PointCoord,
     Barycentric,
+    BarycentricNoPerspective,
     ViewIndex,
 }